requirements.txt
ffmpeg_python==0.2.0
gradio==5.47.2
iou==0.1.0
librosa==0.11.0
matplotlib==3.10.7
numpy==2.2.6
onnx==1.19.1
onnxruntime_gpu==1.23.2
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
Pillow==11.3.0
pyloudnorm==0.1.1
PyYAML==6.0.2
scipy==1.15.3
soundfile==0.12.1
tensorboardx==2.6.4
textgrid==1.6.1
thop==0.1.1.post2209072238
tqdm==4.67.1
transformers==4.46.2
typeguard==4.3.0